global_dev_analytics/requirements.txt
pandas
requests
pyyaml
streamlit
sqlalchemy
openpyxl
pyarrow